fix two typing errors in TCK
1. the type of the count() function is Long in JPA
   (we should clarify this in JDQL b/c it is undefined)
2. we say that only things with the exact same type
   can be compared using =, <>, and friends

+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/AsciiCharacters.java
@@ -90,5 +90,5 @@ public interface AsciiCharacters extends DataRepository<AsciiCharacter, Long>, I
     List<AsciiCharacter> saveAll(List<AsciiCharacter> characters);
 
     @Query("SELECT COUNT(THIS) WHERE numericValue <= 97 AND numericValue >= 74")
-    int twentyFour();
+    long twentyFour();
 }
